Move to ACAP 4 Native SDK and libmodbus version 3.1.11
This patchset moves to the ACAP 4 Native SDK. This means we drop
support for (really) old devices. If you need to build with the ACAP 3
SDK for old devices, please use the previous commit as your base since
that will work fine for that.

With the move to ACAP 4 Native SDK, we can drop the support for this
repo to be built with locally installed SDKs (that nobody uses these
days anyway) and then move the build of libmodbus from Makefile into the
Dockerfile instead.

This patchset also bumps libmodbus from version 3.1.7 to 3.1.11.

Change-Id: I9cfef8b6608c4c8ab897034799b54f8e2aae6fab
Signed-off-by: Joakim Roubert <joakimr@axis.com>